requirements.txt
fire==0.4.0
torch==1.8.1
transformers==4.8.2
pytorch-lightning==1.3.8
bert-score==0.3.11
py-rouge==1.1
nltk==3.6.2
scikit-learn==1.0.2
torchmetrics==0.4.1
